+++ Assets/Scripts/Utility/GeometryHelper.cs
@@ -10,13 +10,13 @@ namespace Utility
             new MyVector2(v.x, v.y);
 
         public static Vector3 ToVector3(this MyVector2 v, float zPos = 0f) => 
-            new Vector3(v.X, v.Y, zPos);
+            new Vector3((float)v.X, (float)v.Y, zPos);
 
         public static Vector2 ToVector2(this MyVector2 v) => 
-            new Vector2(v.X, v.Y);
+            new Vector2((float)v.X, (float)v.Y);
         
         public static MyVector3 ToMyVector3(this MyVector2 v, float zPos = 0f) => 
-            new MyVector3(v.X, v.Y, zPos);
+            new MyVector3((float)v.X, (float)v.Y, zPos);
         
         public static MyVector2 ToMyVector2(this MyVector3 v) => 
             new MyVector2(v.X, v.Y);
@@ -28,8 +28,8 @@ namespace Utility
         {
             MyVector2 p1 = edge.P1;
             MyVector2 p2 = edge.P2;
-            float xMid = (p1.X + p2.X) / 2;
-            float yMid = (p1.Y + p2.Y) / 2;
+            double xMid = (p1.X + p2.X) / 2;
+            double yMid = (p1.Y + p2.Y) / 2;
             return new MyVector2(xMid, yMid);
         }
 
@@ -38,7 +38,7 @@ namespace Utility
             MyVector2 p1 = edge.P1;
             MyVector2 p2 = edge.P2;
             MyVector2 v = p2 - p1;
-            return new Vector3(v.X, v.Y, zPos);
+            return new Vector3((float)v.X, (float)v.Y, zPos);
         }
         
         public static MyVector3 Edge2ToMyV3(this Edge2 edge, float zPos = 0)
@@ -46,7 +46,7 @@ namespace Utility
             MyVector2 p1 = edge.P1;
             MyVector2 p2 = edge.P2;
             MyVector2 v = p2 - p1;
-            return new MyVector3(v.X, v.Y, zPos);
+            return new MyVector3((float)v.X, (float)v.Y, zPos);
         }
 
         public static List<Vector3> ToListV3(this List<MyVector2> thisPolygon)
